services/api-gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
prometheus-client==0.20.0
python-dotenv==1.0.1
